step1: refactor: board -> machine 으로 이름 변경
