+++ src/test/java/lottery/machine/LotteryMachineTest.java
@@ -1,4 +1,4 @@
-package lottery.board;
+package lottery.machine;
 
 import lottery.domain.LotteryRank;
 import org.junit.*;
@@ -9,31 +9,31 @@ import java.util.function.Supplier;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class LotteryGameBoardTest {
+public class LotteryMachineTest {
 
     @Test
     public void test_로또_구매_랜덤번호() {
-        LotteryGameBoard board = new LotteryGameBoard();
+        LotteryMachine machine = new LotteryMachine();
 
-        assertThat(board.buyLotteryTicket(14000)).isEqualTo(14);
+        assertThat(machine.buyLotteryTicket(14000)).isEqualTo(14);
     }
 
     @Test
     public void test_로또_구매_지정번호() {
-        LotteryGameBoard board = new LotteryGameBoard(size -> Arrays.asList(1, 2, 3, 4, 5, 6));
+        LotteryMachine machine = new LotteryMachine(size -> Arrays.asList(1, 2, 3, 4, 5, 6));
 
-        assertThat(board.buyLotteryTicket(14000)).isEqualTo(14);
+        assertThat(machine.buyLotteryTicket(14000)).isEqualTo(14);
     }
 
     @Test
     public void test_로또_구매_지정번호_당첨_1등_2개() {
         Supplier<List<Integer>> winningNumbersSupplier = () -> Arrays.asList(1, 2, 3, 4, 5, 6);
-        LotteryGameBoard board = new LotteryGameBoard(size -> winningNumbersSupplier.get());
+        LotteryMachine machine = new LotteryMachine(size -> winningNumbersSupplier.get());
 
         final List<Integer> winningNumbers = winningNumbersSupplier.get();
-        board.buyLotteryTicket(2000);
+        machine.buyLotteryTicket(2000);
 
-        assertThat(board.checkWinningNumbers(winningNumbers))
+        assertThat(machine.checkWinningNumbers(winningNumbers))
                 .isEqualTo(Arrays.asList(LotteryRank.FIRST, LotteryRank.FIRST));
     }
 }
